requirements.txt
pandas==1.5.0
Pillow==9.4.0
pyspark==3.3.1
requests==2.27.1
scikit_learn==1.2.1
streamlit==1.17.0
streamlit_card==0.0.4
